Support custom comparison in Variants (#11119)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11119

Building on https://github.com/facebookincubator/velox/pull/11021 this adds support
for custom comparison functions provided by custom types to be used in Variants.

This is primarily needed for testing in QueryAssertions.cpp where we convert Vectors
to variants before comparing them.  For example, when a function may choose one of
a set of equivalent values depending on the order of the input (e.g. the representative
value chosen by the Histogram aggregation).

Reviewed By: xiaoxmeng

Differential Revision: D63554289

fbshipit-source-id: 65d627dcbb65d08718bff8b091677905d5b933ee