Refactor Confusing Dwrf Reader RowNumber Method (#11029)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11029

We noticed that the `RowNumber` method in Dwrf reader produces two different outputs based on when it is called. It returns the current position of the reader after a seek operation and then the previous position of the reader before a `next` read operation was called. This makes it almost unusable without an understanding of its underlying implementation. In this diff, we modify it's design to return the current position of the reader for all calls.

Reviewed By: helfman

Differential Revision: D62412793

fbshipit-source-id: cc8036f170b9d3ec2c97858a6932bc9bae14cc0c