Handle missing fields when read flatmap as struct (#11106)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11106

When we read flatmap as struct, the presence of a field is not decided
the same way as a normal struct field.  We need to handle this separately and
set `ScanSpec::subscript` according to the flatmap specific logic, and also
leverage this subscript in the subsequent reading to treat it as constant.

Reviewed By: pedroerp

Differential Revision: D63469632

fbshipit-source-id: 067c3f9fc82c95d4e0ad667ca20a174f343abc99