Add basic group by and status returning without unified page faults (#10944)
Summary:
Add basic group by

Adds an interpreted group by with only bigint summs and single bigint
key.  Adds a continue mechanism used when the table runs out of space
in mid kernel. Failed lanes are restarted after filling up the device
side allocator and possibly rehashing the hash table.

Adds a rehash function and logic for allocating group by rows as
needed. The sizes are set up so that the rows to allocate run out
before the slots in the table, so that only the availability of new
rows needs to be checked.

- Sets the memory order for keys to be release on write and consume on
  read. Write the first key last and read the first key first to order
  other stores with the first key acting as the atomic for memory
  order.

- Adds a Task-wide barrier for Wave for cases of updating resources
  (hash tables) shared between Driver pipelines.

- Next step is to use the code as a a codegen template for different
  key/aggregate combinations.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10944

Reviewed By: Yuhta

Differential Revision: D62333042

Pulled By: oerling

fbshipit-source-id: b03d01c3c150a5193ad2bf1d2bcdc58df1ad4abc