Support Custom Comparison in Simple Function API (#11136)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11136

This change extends CustomType in the Simple Function API, templating it on
providesCustomComparison, with the idea that Custom Types that provide custom
comparison operators will set this value to true when they define their CustomType
template overrides.

In Simple Functions, this allows us to wrap the value in a view, called
CustomTypeWithCustomComparisonView, which overrides the standard comparison
operators and hash function to use the operators provided by the Type.  It also provides a
dereference operator to access the raw underlying value.

The only other piece to this diff is adding a VectorReader that constructs the
CustomTypeWithCustomComparisonView.

This should help address https://github.com/facebookincubator/velox/discussions/8713
so we don't have to create special flavors of UDF just for TimestampWithTimezone.

On a side note, if we extend this so that if a Custom Type does not provide custom
comparison operators we produce a similar view without the custom comparison operators,
that would likely solve the issue with custom types and the simple function interface
entirely.

In the follow up diff, I'll clean up the TimestampWithTimezone UDFs for comparisons and
greatest/least.

Reviewed By: xiaoxmeng

Differential Revision: D63670301

fbshipit-source-id: dc94e86c6caa645f18f8e84c1f84eb514af6e648