Support custom comparison in Histogram Aggregation (attempt 2) (#11154)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11154

This is a second attempt to land the changes in
https://github.com/facebookincubator/velox/pull/11120

The original description:
Building on https://github.com/facebookincubator/velox/pull/11021 this adds support for custom
comparison functions provided by custom types in the Histogram aggregationt.

New context:
I landed this along with https://github.com/facebookincubator/velox/pull/11119 so it
got reverted along with it.  This particular change did not introduce any issues
though.

Reviewed By: xiaoxmeng

Differential Revision: D63795479

fbshipit-source-id: 6c045c8648abe600e5656d3874c47aba21186168