Add option for admission control for filesystem resources (#11146)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11146

This change adds a generic admission controller class that can be
used for filesystem resources like read bytes in flight or number
of read requests in flight. It also provides a way to report stats
for resource usage, queued count, queued wait times by allowing the
client to specify a metric name.

Reviewed By: xiaoxmeng

Differential Revision: D63728285

fbshipit-source-id: b880549ecaaed727ff4c459202a101fd71a00d8d