Make exceededMaxSpillLevelLimit_ in HashBuild and HashProbe tsan_atomic (#11165)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11165

setupSpiller and canReclaim may be called concurrently in HashBuild and HashProbe
which leads to TSAN errors in our tests as the write to and read from
exceededMaxSpillLevelLimit_ respectively.

In practice reclaim checks for the case where exceededMaxSpillLevelLimit_ has been
updated since the last time canReclaim has been called and handles that case, so this
will not cause issues.

Marking exceededMaxSpillLevelLimit_ tsan_atomic to get the tests green.

Reviewed By: xiaoxmeng, tanjialiang

Differential Revision: D63853455

fbshipit-source-id: a633f92fe521004b7b99a26b8fea29647b7e29c3