Support custom comparison in Map Aggregations (attempt 2) (#11155)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11155

This is a second attempt to land the changes in
https://github.com/facebookincubator/velox/pull/11124

The original description:
Building on https://github.com/facebookincubator/velox/pull/11021 this adds support
for custom comparison functions provided by custom types in the map aggregations
map_agg and map_union.

Note that I did not add support in map_union_sum because it currently only supports
Maps with a fixed set of key types, and none of those provide custom comparisons.

New context:
I landed this along with https://github.com/facebookincubator/velox/pull/11119 so it
got reverted along with it.  This particular change did not introduce any issues
though.

Reviewed By: xiaoxmeng

Differential Revision: D63796041

fbshipit-source-id: 06e7ed5e5f62e8671341c6b938429aa05b23c177