Add TableWriterRepalyer (#11100)
Summary:
Adds `TableWriterReplayer` to facilitate the replaying of `TableWriter` operator.
Uses the given plan node ID to find the traced `TableWriteNode` from the traced plan.
It helps create a new `TableWriterNode` and rebuild a query plan with a `QueryTraceScanNode`,
then apply the traced configurations, and rerun.

`QueryTraceScanNode` holds the traced data type and dir for a given plan node ID.
These information can be utilized to build the `QueryTraceScan` operator. It creates
a `QueryDataReader` using the traced data type and input data file. To find the right
input data file for replaying, we need to use both the pipeline ID and driver ID, which
are only known during operator creation, so we need to figure out the input traced
data file and the output type dynamically.

Part of https://github.com/facebookincubator/velox/issues/9668

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11100

Reviewed By: tanjialiang

Differential Revision: D63774083

Pulled By: xiaoxmeng

fbshipit-source-id: 912bef3cb20d9b1a1685af625ba2f319e2dc7509