Fix data race in cache eviction path (#11193)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11193

The cache entry init code path (thread) logs an error message with cache entry size when allocation fails.
This happens after it releases the cache entry which could be race with cache eviction code path which clears
to zero. The fix is to print out the allocated pages instead of bytes in cache error message to prevent.

Reviewed By: kevinwilfong, zacw7

Differential Revision: D64014938

fbshipit-source-id: d752c6bf13e404c9a3553dd548795ec5534e961b