Fix Presto serde bug when deserializing large payload (#11177)
Summary:
codec gives back chained iobufs when payload is very large. We only take care of unchained scenario and was able to get away with it. Fix the case for large payload and added unit test for it. The estimated buffer size to trigger this bug is around 80MB in GZIP, ZSTD and ZLIB

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11177

Reviewed By: xiaoxmeng

Differential Revision: D63962642

Pulled By: tanjialiang

fbshipit-source-id: cbaf2bb5518de786c69461b5f8d725732c9f6fe8