Add CODEOWNERS file (#11081)
Summary:
The CODEOWNERS file allows us to setup file based rules that will automatically request a review from owners on PRs with changes to matching files.

We currently do not enforce these reviews as required so it's only a tool for more granular notifications at the moment. For example component maintainers can set a rule so that they are pinged on changes to the sections of the codebase that are relevant for their component.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11081

Reviewed By: bikramSingh91

Differential Revision: D64055409

Pulled By: xiaoxmeng

fbshipit-source-id: 94e168bfbd71252e57f04082cdb1b876c6f9c696