Fix arrow install with custom INSTALL_PREFIX (#11052)
Summary:
If the recently added INSTALL_PREFIX does not point to /usr/local in Linux then the bundled Thrift installation from the Arrow dependency fails because it cannot locate Boost.

Boost was previously installed into the INSTALL_PREFIX but the Thrift CMake in Arrow does not pass on any PREFIX_PATH settings to Thrift. The PREFIX_PATH is set to the INSTALL_PREFIX and works with Arrow itself just fine.

The error would be

Could NOT find Boost (missing: Boost_INCLUDE_DIR) (Required is at least
  version "1.56")

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11052

Reviewed By: bikramSingh91

Differential Revision: D64060501

Pulled By: xiaoxmeng

fbshipit-source-id: b393d9d0590828a8473a3f80ebc10d350fd3d20a