Avoid additional memory allocation during sort output spill (#11199)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11199

Memory arbitration fuzzer is flaky because of the unexpected memory allocation after sort output spill finish.
The reason is that when we finish spill, we setup the merge reader to prepare reading the unspilled data.
The merge reader might use non-trivial amount of data which cause additional memory consumption. And
we shall keep the spill critical path as fast as possible. This PR moves the merge reader setup from the spill
path to the first get output with unit tests. Also improve the fuzzer test logging a bit to help debug.

Reviewed By: bikramSingh91

Differential Revision: D64072660

fbshipit-source-id: 01e409a92a329b1634cf746af940a9e2b8af11c4