Add clean shutdown for ExchangeFuzzer runs (#11191)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11191

After the ExchangeFuzzer runs all the Tasks need to shutdown, and as part of
shutting down they have a future to abort any Tasks still running as they should all
terminate as a unit.  There's a chance not all Tasks have completely finished by the
time the function terminates, i.e. this future hasn't run yet. This can lead to TSAN
issues as there's a race between when the Tasks are deleted and these futures. To
ensure everything happens in the right order, I've added a clean shutdown step at the
end of ExchangeFuzzer runs just to make sure that all the Tasks have finished.

Reviewed By: xiaoxmeng

Differential Revision: D64010841

fbshipit-source-id: cac51eab3d37aea55ab4a1c19c6781005ecc9b8a