Ignore tracing of auxiliary operator (#11220)
Summary:
Ignore the tracing of auxiliary operators of a traced operator within
the same driver, sharing the same node ID to prevent disruptions.
For instance, consider the `TableWriter` operator, which has an auxiliary
`Aggregation` operator. Both share the same node ID and operator ID,
which causes them to initiate tracing sequentially, and create the same
 tracing data directory using the same node ID and driver ID leading to
disruptions and conflicts.

This PR introduces a tracing operator map per driver context. It stores the
operator ID to the operator type preventing the auxiliary operator from creating
a tracer. The tracing processes and directories for operators with the same
node ID are isolated between drivers so the map scoped a single driver.

part of https://github.com/facebookincubator/velox/issues/9668

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11220

Reviewed By: tanjialiang

Differential Revision: D64223496

Pulled By: xiaoxmeng

fbshipit-source-id: 20a641ebec6cbe72c9a31e1345398c9d0bdef0e1