Add tests constructing maps with duplicate custom TimestampWithTimezones (#11228)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11228

When constructing a map in Presto and checking for duplicate keys, it already uses
the key Vector's equalValuesAt function which calls the custom comparison operator
provided by the custom type (if that applies).  So this just works automatically for
types with custom comparison.

Adding unit tests to verify this and ensure it doesn't break in the future.

Reviewed By: xiaoxmeng

Differential Revision: D64214359

fbshipit-source-id: 15261dd7cf32693a199eba9b7efc43f47c39177b