Add tests for Presto's map_top_n/map_top_n_keys UDFs with TimestampWithTimezones (#11243)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11243

Presto's map_top_n and map_top_n_keys implementations use the hash and equals
functions implemented for GenericViews.  For TimestampWithTImeZone this
recursively calls these functions on the CustomTypeWithCustomComparisonView the
GenericView wraps, which supports custom comparison operators. So these UDFs
work as is with the recent changes earlier in this stack.

Adding unit tests to verify this and ensure it doesn't break in the future.

Reviewed By: amitkdutta

Differential Revision: D64267032

fbshipit-source-id: b7fdd725cae5918c98a9cb0a6d053966991f00c5