Remove unused-variable in velox/common/memory/MemoryAllocator.cpp (#11251)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11251

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: meyering

Differential Revision: D64278704

fbshipit-source-id: 93b8a43f6fc15ff04ceae712f3b0e166ca9ea5a1