Remove unused-variable in velox/dwio/dwrf/reader/SelectiveStringDictionaryColumnReader.cpp (#11274)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11274

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: meyering

Differential Revision: D64279128

fbshipit-source-id: d0ce7f111be62ac3a6a06ea2c4d7025b4ba5257d