Add shutdown support memory arbitrator (#11325)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11325

There are bunch of flaky mock arbitrator tests which are due to the background global arbitration
which can continue to run after the test finishes.
In general we need the shutdown procedure to handle the shared arbitrator destruction properly.
This PR adds shutdown support in shared arbitrator and is invoked from memory manager. This fixes
plus some test fixes solves all existing flakiness in mock arbitrator tests in Meta.

Reviewed By: tanjialiang, kagamiori

Differential Revision: D64742451

fbshipit-source-id: 84d202f4e6926be030b009b2a1128372b1f619a5