Allow InternalPlanNodes to accept a PlanVisitor
Previously, we would fail if an InternalPlanNode tried to accept a
PlanVisitor that was not an InternalPlanVisitor. This was useful to tell
developers to use InternalPlanVisitor when operating in presto-main.
However, it prevented plugins from being able to traverse a plan
that contains InternalPlanNodes. Allowing plugins to traverse the
plan has becoeme critical since adding support for plugin
PlanCheckers. Without this change, plugin plan checkers become almost
useless.