Avoid local shuffle and global shuffle use the same hash value to avoid data skew (#11338)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11338

There is data skew problem in Velox partitioning handling for local and remote shuffle used by Prestissimo.
If the number of workers for remote shuffle is multiple of the number of threads in local shuffle. Then all
the input on a worker could goes to the same thread. For system hash partitioning, it causes overall slow in
query execution time. For hive partitioning used by table writer, it could also cause exceeded open file
errors as the limit is per table writer thread. We shall use a different hash for the local shuffle. For hash
partition, we apply a second hash function on the computed hash number for the local shuffle while
for the hive partition, we shuffle the bucket to partition map.

Verified in shadow testing, this can prevent exceeded open file errors for hive partition use case. And
significantly improve the query execution time with the default hash partition setting in Prestissimo.

Reviewed By: kevinwilfong, arhimondr, oerling

Differential Revision: D64777771

fbshipit-source-id: 0b54cb1c2fa64d67d02f5ba39c3d839ffec4cb6a