Propagate driver context in parallel join build (#11329)
Summary:
When hash build is building join table, it suspends the current running driver, to prevent the parallel building threads from triggering arbitration and create a deadlock situation. The downside of this approach is, we lose the opportunity of reclaiming from it, because we will not wait for the table build to finish and become reclaimable when reclaiming from the operators.
This PR removes the suspension logic. Instead it pushes in the driver context to the parallel building threads so that these threads, when triggering arbitration (because suspension is removed), can correctly suspend the current running driver, so that the reclaiming can wait for it to finish the table build and get a chance to reclaim from it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11329

Reviewed By: xiaoxmeng

Differential Revision: D64795445

Pulled By: tanjialiang

fbshipit-source-id: 3856645ede8acc82690c3a28aecec6e1c61b560c