Let HashJoinBridge be able to reclaim memory in the middle state (#11324)
Summary:
Add ability to allow hash join bridge to be able to reclaim directly the transferred table from hash build. This is useful in the condition when build has finished and table has been transferred to bridge, but probe side was blocked by downstream operators from performing any of its initial isBlocked() calls. When this happens, probe does not have the ability to spill because it has no knowledge about the ongoing join, neither does build because the table is transferred to bridge already. So when this happens, we rely on bridge to perform the table spill.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11324

Reviewed By: xiaoxmeng

Differential Revision: D64785513

Pulled By: tanjialiang

fbshipit-source-id: a423e55bd251c1dd4256998548be8961c8689f57