Fix the global arbitration check failure when it is disabled (#11364)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11364

Some memory intensive queries run into check failure on global arbitration flag before they
start to wait for global memory arbitration as it is not enabled in those clusters. The issue is because
of a bug that we fall back to global arbitration when query which fails to allocate memory from the system
and within the capacity limit and is not qualified to do memory reclaim from itself.
This PR fixes this with unit test.

Reviewed By: tanjialiang, arhimondr

Differential Revision: D65073054

fbshipit-source-id: 85eb4d5a3102fde4d8989ea5a0f076b2bc2c04d8