Fix tsan flaky test MockSharedArbitrationTest.concurrentArbitrationWithTransientRoots (#11365)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11365

concurrentArbitrationWithTransientRoots detects a data race when we try to log root pool's capacity
from a leaf pool which doesn't hold the lock on the root but only the leaf. This is mostly for logging and
we don't want to introduce the locking on the root which might cause potential deadlock or performance
issue. Change this to tsan variable to pass the test.

Reviewed By: tanjialiang

Differential Revision: D65075928

fbshipit-source-id: 88ad38cdcb1a4f991c224e99080020d4448eb32f