Add native plan checker to sidecar plugin
This adds a provider for a native plan checker that will send a plan fragment
to the native sidecar where it is validated by performing a conversion to a
Velox plan. If the conversion succeeds the query will continue, if it fails
then the query will fail with an error from the native sidecar. The provider
is added to the native sidecar plugin and is enabled with the config
`native-plan-checker.plan-validation-enabled=true` from filename
`etc/plan-checker-providers/native-plan-checker.properties`.

See also: prestodb#23649
RFC: https://github.com/prestodb/rfcs/blob/main/RFC-0008-plan-checker.md