Add Row ID column reading support (#11363)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11363

Create framework in selective readers for generating row number field
and composite with other constant fields.  Then use this framework to add
support for Row ID column.  We move the row number generation logic into column
reader during this process, so that we could filter them if needed in the
future.
bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D65028172

fbshipit-source-id: 716dbdc77301267366a10302459e105d711a92f0