Support expression transformers in expression fuzzer (#11264)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11264

Some Velox functions returns Array-typed results with different order
of elements from Presto, such as array_intersect, array_except, etc.
This mismatch of element order is cosmetic because the order is not
defined. To tolerate this mismatch in fuzzer, this diff introduces an
SortArrayTransformer that wraps these array functions in array_sort
when generating the random expressions.

There is a limitation with the current SortArrayTransformer that it
doesn't support arrays that contain maps because map is not an
orderable type. In this situation, SortArrayTransformer transform the
expression into a NULL constant. This can be extended later to apply
`array_sort(a, (x, y) -> if(json_format(cast(x as json)) < json_format(cast(y as json)), 1, if(json_format(cast(x as json)) = json_format(cast(y as json)), 0, -1)))`
if the array contains map.

Reviewed By: bikramSingh91

Differential Revision: D64358222

fbshipit-source-id: 12b35def70b1fb8297499a09324ad3292b3052c5