Fix flakiness in TraceUtilTest.getTaskIds (#11394)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11394

Currently the test verifies the contents of a vector that fetches
filenames from a directory. However, the filesystem call to fetch
the contents of a directory might not always be in a sorted order
which can fail the verification step of this test. Therefore, this
change ensures the contents are sorted before verification.

Reviewed By: xiaoxmeng

Differential Revision: D65244642

fbshipit-source-id: 46ce3f59cd08bb18f4e21618e81551dd1aed5721