Fix flaky test MockSharedArbitrationTest.arbitrateBySelfMemoryReclaim (#11397)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11397

The test originally intended to test local arbitration. But after refactoring, this test is by default relying on global arbitration for reclaiming. Because global arbitration runs in a different thread, there is a race condition between the actual reclaim and the reclaimed bytes check in the test. Making the test disable global arbitration solves the flakiness.

Reviewed By: xiaoxmeng

Differential Revision: D65259718

fbshipit-source-id: 86a2937c620d1ac58bd9b76c9948f9d9b0aaa0d5