Remove unused-variable in velox/buffer/tests/BufferTest.cpp (#11398)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11398

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: aaronwlma, pedroerp

Differential Revision: D65144811

fbshipit-source-id: 79812cf160aaeeec514885c9a2bbfd860c87c1c7