Fix HiveTypeSerializer.unregisteredOpaque (#11401)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11401

Currently the opaque types are registed into a global static variable that is shared
across the process. If multiple tests within the same Suite use the same alias,
they can interfere with the test's expectations. Therefore, this introduces a way
to unregister the type to allow isolation between tests.

Reviewed By: kunigami

Differential Revision: D65290938

fbshipit-source-id: 9ce33ffcb956b7195ec0e1e9eb32c981bac57eb9