Use fallocate for file size extension when supported (#11403)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11403

When Copy-on-Write (COW) is disabled on Btrfs, automatic relocation
creates snapshots of files, ignoring the noCOW setting. This results
in increased disk usage and can lead to "no space left" errors in
production.

One possible enhancement we can make is to use `fallocate` to reserve
space immediately after file creation. This helps ensure the allocated
space is as continuous as possible.

Reviewed By: xiaoxmeng

Differential Revision: D65316028

fbshipit-source-id: f690fb454d40b45c005358c85d5389b75e7ebe77