Remove unused-variable in velox/buffer/tests/BufferTest.cpp +8 (#11421)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11421

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D65282800

fbshipit-source-id: 3ebdc85005aa8a250db6c5e72826d18a7a95f926