Respect kPreferredOutputBatchRows strictly in Unnest (#10711)
Summary:
Before this patch, it has restriction that one row output should be in one batch, an unusually long array or map in a single row may result in a very large batch size, which can cause OOM when later processed by column to UnsafeRow operator  because row format may cost more memory even to 10x than column format for nested struct column type.
After this patch, it is split to several batches to match output batch size kPreferedOutputBatchRows.
Record the start of the first row, process the first row and last row in different branch, the middle rows range is 0 to maxSize while the first row range is firstRowStart_ to maxSize and last row range is 0 to maxSize or lastRowEnd depending on if the lastRowEnd is available when there is not only 1 row to process.
To cover almost all the cases, the Unnest test is changed to parameterized test which accepts several batch size {2, 17, 33, 1024}, even more, I test batch size {2, 5, 7, 12, 16, 17, 33, 45, 50, 60, 71, 1024} local,  it does not appeal to any more problem, so current approach is valid enough.

Resolves https://github.com/facebookincubator/velox/issues/10655

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10711

Reviewed By: Yuhta

Differential Revision: D65473321

Pulled By: mbasmanova

fbshipit-source-id: 4038a286d49c7a83f7e5825f4c9d232a27a11a09