Add gcc11 to Ubuntu20.04 setup and add PkgConfig install (#11326)
Summary:
Modernize the compiler used on Ubuntu 20.04. Once Velox moves to using the C++20 standard the system gcc9 compiler cannot be used anymore.

In addition, recent upgrades added the usage of PkgConfig which was not installed into the system and is not by default pre-installed into ubuntu 20.04.

Resolves: https://github.com/facebookincubator/velox/issues/10953

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11326

Reviewed By: Yuhta

Differential Revision: D65420229

Pulled By: mbasmanova

fbshipit-source-id: 48e6108d1891b85ed9f4d06dc81fa851076c7748