Add support for S3 bucket config (#11321)
Summary:
Allow all `hive.s3` options to be set on a per-bucket basis.
The bucket-specific option is set by replacing the hive.s3. prefix on an option
with hive.s3.bucket.BUCKETNAME., where BUCKETNAME is the name of the bucket.
When connecting to a bucket, all options explicitly set will override the base hive.s3. values.
These semantics are similar to the Apache Hadoop-Aws module.
https://hadoop.apache.org/docs/current/hadoop-aws/tools/hadoop-aws/index.html
Spark uses this for ETL workloads between 2 buckets.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11321

Reviewed By: Yuhta

Differential Revision: D65365858

Pulled By: kgpai

fbshipit-source-id: 0a0cf401ef22d7270da7e932a99c0d8614597299