[codemod] Remove unused-variable in gen_ai/unified_parallelism/up/parallelisms/runtime/collective/MNSendRecv.cpp (#23989)
Summary:
Pull Request resolved: https://github.com/prestodb/presto/pull/23989

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D65584210