Add option to disable flatenning optimization in PrestoSerializer (#11465)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11465

This change introduces an option, "preserveEncodings" to the SerDe
Options, to disable the flattening optimization in the
PrestoBatchVectorSerializer, enabling support for use-cases
where encodings need to be preserved between the serialization
and deserialization steps.

Note:
There are still 2 cases where encodings can be lost:
1. Nulls in dictionary layer: Presto page format does not support having
nulls so this gets flattened
2. If there are multiple levels of dictionary or a constant underneath:
Currently, the presto serializer only preserves the top most layer and
flattens everything underneath it.

Reviewed By: kevinwilfong, kagamiori

Differential Revision: D65569946

fbshipit-source-id: 5e38dce14d8741b504d8adc6159551d89af663c9