Remove unused-variable in velox/experimental/wave/exec/Wave.cpp (#11484)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11484

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D65584300

fbshipit-source-id: 7bd71c4f786ff7e35aa5a6650f1cb260f996163e