Remove unused-variable in velox/common/memory/tests/MockSharedArbitratorTest.cpp (#11485)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11485

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D65584421

fbshipit-source-id: 5609bf5fef8ce066201f54f0457467143e2ce5f1