Add TableScan Replayer (#11205)
Summary:
Add a split tracer to record and load the input splits from a tracing 'TableScan'
operator, and for getting the traced splits when replaying 'TableScan'.
Currently, it only works with 'HiveConnectorSplit'. In the future, it will
be extended to handle more types of splits, such as 'IcebergHiveConnectorSplit'.

part of https://github.com/facebookincubator/velox/issues/9668

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11205

Reviewed By: tanjialiang

Differential Revision: D64946986

Pulled By: xiaoxmeng

fbshipit-source-id: 5f656e5ad9b8755484c3eb5ac42b209569a98101