CMake: Move resolve dependency macros under velox_ prefix (#11392)
Summary:
It is best practice in CMake to use a project prefix for project specific macros, functions and variables. This makes it easier to compose projects in larger builds as well as differentiate general functions from project specific functions.

This moves several macros and it will require some simple migration of any CMake code reusing the resolve dependency macros:

 * build_dependency -> velox_build_dependency
 * list_subdirs -> removed after reviewer suggestion
 * resolve_dependency -> velox_resolve_dependency
 * resolve_dependency_url -> velox_resolve_dependency_url
 * set_source -> velox_set_source
 * set_with_default -> velox_set_with_default

This should be a simple change as the other changes are just to adapt to the changed macro names.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11392

Reviewed By: xiaoxmeng

Differential Revision: D65763456

Pulled By: kgpai

fbshipit-source-id: f748e5431e246dc8c095c4838ef2f4b96de6acf2