Remove unused-variable in turbine/state_syncer/StylusCommonActionApi.cpp +3 (#11505)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11505

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D65755217

fbshipit-source-id: 77d4f0046fdee8f833ec1f989db4d2a1939a5b3e