Remove unused-variable in velox/experimental/wave/exec/ToWave.cpp +3 (#11504)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11504

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Differential Revision: D65755246

fbshipit-source-id: 8dfce093a9c638f0630fe7460a029f266083c834