Ensure prefix length does not exceed configured maximum (#11496)
Summary:
The check can't make sure the normalizedKeySize is less than or equal to
maxNormalizedKeySize. That is possible
`normalizedKeySize + encodedSize.value() > maxNormalizedKeySize` but related
column is included in the prefix.
Move the check to fix this.
This is found when doing the PR: https://github.com/facebookincubator/velox/pull/11272

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11496

Reviewed By: tanjialiang

Differential Revision: D65850949

Pulled By: xiaoxmeng

fbshipit-source-id: 2dc5fcb93f0960544b9df3f2471caa2923c70975