Remove unused-variable in velox/dwio/parquet/reader/DeltaByteArrayDecoder.h (#11528)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11528

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: meyering

Differential Revision: D65859977

fbshipit-source-id: 96e0985283e7fe607c80217dcd1828aa6ac79e01