Allow single thread execution mode for local partition to support spark union for gluten (#11531)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11531

Local partition could be used to support spark union for gluten. For union, it will be translated to local partition
with a followup distinct aggregation. Gluten is single thread execution mode using API like task::next to pull data.
Currently, Velox only allow single thread execution mode if the front of a pipeline is not local exchange which is not necessary.
Task::next actually pull results from all the pipelines so it should be good. This PR removes that check and verified with the
unit test.

Reviewed By: bikramSingh91, oerling

Differential Revision: D65917749

fbshipit-source-id: 1ebe8c5782d7d36d4a77cdc792169856c03d0dbd