Combine low selectivity vectors generated by the hash join filter (#10987)
Summary:
Combine low selectivity vectors generated by join filter
**Problem**
We found in TPCDS query72 that the join filter leads to a large number of
low-selectivity result vectors, which affects the performance of subsequent
operations.

Details:
The number of input vectors on the probe side of the corresponding join is
84,087, with a total of 743,851,486 rows (our batch size is set to 10,240). Due
to a large number of duplicate rows on the build side, the final result
inflates. The number of output vectors from the join is 2,806,054. The
corresponding join filter filters out a large portion of the results, so the
 number of output rows is 1,430,253,235. This leads to the output of many
 sparse vectors (the average batch row count is 504).

 **Solution**
The original logic continues the [loop](https://github.com/facebookincubator/velox/blob/main/velox/exec/HashProbe.cpp#L1065) to fill more rows if no rows pass the
filters. To resolve this issue, we can extend it to handle cases where only
partial rows pass the filters. We need to ensure that the indices
'outputRowMapping_' and 'outputTableRows_' are filled as much as possible until
we either reach the preferred batch size or have processed all rows in the
current input vector.

This approach will not only address the issue mentioned above but also avoid
unnecessary data copying.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10987

Reviewed By: pedroerp

Differential Revision: D65949331

Pulled By: Yuhta

fbshipit-source-id: e8e54e7be1b5a93ba71387aa49f00dcff3cc8aa2