build(ci): Add conventional commit title check (#11520)
Summary:
Check the PR title against cc pattern added  in https://github.com/facebookincubator/velox/issues/11356. Fails with a pointer to  CONTRIBUTING.md.
We are running the title change on every commit as it takes under 5 seconds to run and this way we avoid weird states with skipped checks overriding failed results when the title was triggered outside of a commit.

We could add a comment to the PR but that would require a workflow with elevated permissions using `pull_request_target` which should be avoided if possible. So I think the marginally lower ux is justified, we still get a marker/reminder for reviewers through this job.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11520

Reviewed By: pedroerp

Differential Revision: D65947963

Pulled By: Yuhta

fbshipit-source-id: 290b97f784011367d3e481ff49f42b7da65fdf7a