- Fix FlatVector copy memory issue (#11483)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11483

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11482

When copy-ing a FlatVector the source memory pool is used. This is a problem when the vector needs to live beyond the lifetime of the source pool.

For strings this requires recomputing new StringViews over newly copied string buffers. This is useful in cases where you need to maintain result rows beyond the lifetime of a given presto task / query.

Reviewed By: xiaoxmeng, arhimondr

Differential Revision: D65306907

fbshipit-source-id: 2ca69a808131d97f78483166fa5736ca0f4a4034