[native] Add config for cpp cache default behavior
In Presto Java, cache is disabled by default and only enabled when the config `node_preference` is set to `SOFT_AFFINITY`, while in CPP, cache is enabled by default and only disabled when the config node_preference is set to "NO_PREFERENCE".

This change adds a new config to support the alignment of the default cache behavior.