feat: Use fallocate for file size extension when supported (#11541)
Summary:
When Copy-on-Write (COW) is disabled on Btrfs, automatic relocation creates snapshots of files, ignoring the noCOW setting. This results in increased disk usage and can lead to "no space left" errors in production.

One possible enhancement we can make is to use fallocate to reserve space immediately after file creation. This helps ensure the allocated space is as continuous as possible.

For environments where "no space left" errors are already happening, attempting to reserve space can result in failures. In that case, fallback to truncate and record it in runtime metrics.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11541

Reviewed By: xiaoxmeng

Differential Revision: D65977522

Pulled By: zacw7

fbshipit-source-id: 6bed41cf6ce5d67327cf1820c4589c8f89262fab