fix: Fix Presto Java UUID serialization (#11197)
Summary:
This fixes the `PrestoSerializer` to put UUID values in the correct format that is expected by Presto Java so that the values will match those from a Java worker. First, when converting UUID to/from string, the values are no longer in big endian format (as taken from boost::uuid) and are instead stored as a little endian in an int128_t. Secondly, Presto Java will read UUID values from an `Int128ArrayBlock` with the first value as the most significant bits. To correct this, the upper/lower parts of the int128_t are swapped during serialization/deserialization.

A unit test for checking roundtrip UUID serializaiton was added and manual testing of Presto with a native worker to verify the problem from the issue description is fixed.

From https://github.com/prestodb/presto/issues/23311

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11197

Reviewed By: Yuhta

Differential Revision: D66204014

Pulled By: kagamiori

fbshipit-source-id: 1e2842df57e672d6cf056a658108691e3ad39d15