fix: Match Presto's behavior for invalid UTF-8 in url_encode (#11518)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11518

Presto Java converts the URL to a Java String before encoding it in url_encode. Java replaces bytes
in an invalid UTF-8 character with 0xEF 0xBF 0xBD.

Velox encodes invalid UTF-8 characters as is, which leads to differences in results from Java and
C++.

This diff adds a check when encoding URLs for invalid UTF-8 characters and does the same
replacement as Java.

Reviewed By: kgpai

Differential Revision: D65856104

fbshipit-source-id: c661aa0c365bf7042d822ff7c7186358b1e595c7