fix: Match Presto's behavior for invalid UTF-8 in url_decode (#11604)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11604

Presto Java converts the URL to a Java String after decoding it in url_decode. Java replaces bytes
in an invalid UTF-8 character with 0xEF 0xBF 0xBD.

Velox decodes invalid UTF-8 characters as is, which leads to differences in results from Java and
C++.

This diff adds a check when decoding URLs for invalid UTF-8 characters and does the same
replacement as Java.

Reviewed By: xiaoxmeng

Differential Revision: D66249265

fbshipit-source-id: d6993f8c70e72368b85b4c64814a45a1359f0802