fix: Handle escaped separators in Presto url_extract_parameter function (#11540)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11540

Presto Java supports the case where a separator ('=' or '&') is percent escaped in the query portion
of the URL in url_extract_parameter.

Velox currently only unescapes the value of the parameter once it's been found, so when a separator
is escaped it breaks parsing the parameters in the query.

This change fixes it so Velox's url_extract_parameter unescapes the query portion of the URL before searching for the parameter in it.

Reviewed By: xiaoxmeng

Differential Revision: D65967602

fbshipit-source-id: bbbdfbd40d5598a71e0adde5a3bb8d8a725956fe