fix: Fix hash join reclaim bytes accounting (#11624)
Summary:
Current hash join reclaim bytes accounting is wrong. It let the later reclaimed op type override the previous one, which we should be adding. This made some reclaimed bytes from hash join to be 0 because when it firstly reclaims from build it might reclaimed some bytes but later when it tries to reclaim from probe, it reclaimed 0 because probe has not started yet. Then the reclaimed result is 0 which is wrong. This wrong reclaimed bytes can make the reclaiming framework to reclaim from additional nodes which could at most of the times be avoided. For some hand picked spilling queries, fixing this can bring up to 50% of latency gain.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11624

Reviewed By: xiaoxmeng

Differential Revision: D66381505

Pulled By: tanjialiang

fbshipit-source-id: d88d2bc110505a11335443f437e3c3b12f72f5d7