fix: Move the thread local parameter from method to member parameter in HdfsReadFile (#11580)
Summary:
To address this [comment](https://github.com/facebookincubator/velox/pull/9835#discussion_r1775374914), we moved the thread local parameter `folly::ThreadLocal<HdfsFile> file_` from a member parameter to a template parameter in the `HdfsReadFile::preadInternal` method. The frequent initialization of the thread local parameter in the method caused about a 5% performance degradation. We reverted the thread local parameter to a member parameter in this PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11580

Reviewed By: kgpai

Differential Revision: D66376694

Pulled By: kagamiori

fbshipit-source-id: ba639dd9bd52a81a71d2580d3af5259324a8c92a