feat: Add skewed partition balancer (#11635)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11635

Add skewed partition balancer which implements the following functions:
(1) the task assignments for a number of partitions;
(2) tracks the processed bytes/rows for each partition and the corresponding estimates of task and partition workloads in terms
of the processed bytes;
(3) rebalance (or scaling busy partition processing) by assigning under-loaded tasks to hot partition from a over-loaded tasks.
(4) get the next serving task for a given partition in round-robin manner if there is multiple assigned tasks.

The followup is to leverage this to auto-scale writer processing for hot table partition through a customized local partition for
table writer

Reviewed By: tanjialiang, oerling

Differential Revision: D66380972

fbshipit-source-id: cda307f48bbc362e8cdaa79ff0cb427f99d954c3