fix:Support hashing ipaddress (#11640)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11640

Checksuming on ipaddress fails because the prestoHasher treats ipaddress as a HUGEINT, which it is; however, the hashing for HUGEINT is different than the hashing for binaries.

In Velox, we represent IPAddress as a HUGEINT (16 bytes) instead of varbinary. So we need to compute the checksum on the varbinary representation of the int128_t.

Reviewed By: bikramSingh91

Differential Revision: D66416538

fbshipit-source-id: 6becaae50c9cc09132190b65811f068b4578be48