fix: Fix Stream not close because it holds in TreeOfLoser (#11660)
Summary:
SpillPartion::createOrderedReader opens all the spill files and holds it in TreeOfLoser, SpillFile -> FileInputStream->buffer_(readBufferSize), so it holds memory readBufferSize * numFiles,  so we need to close the stream as soon as possible after it reaches the end as Spark does.
After close, the function SpillMergeStream::hasData() returns false, then it won't access the fields of SpillMergeStream.
https://github.com/apache/spark/blob/master/core/src/main/java/org/apache/spark/util/collection/unsafe/sort/UnsafeSorterSpillReader.java#L134

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11660

Reviewed By: bikramSingh91, tanjialiang

Differential Revision: D66658505

Pulled By: xiaoxmeng

fbshipit-source-id: 7a4d7c158cb688ff947585d36f59b44816351ba1