feat(planbuilder): Enable passing compressionKind via tableWrite (#11724)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11724

Currently there is no way for users to specify `compressionKind` in the velox plan builder in order for the output to be compressed using one of the supported schemes defined by the `CompressionKind` enum. Instead, we always use the hard-coded `CompressionKind_NONE` value.

This diff fixes this by exposing a new `compressionKind` option in one of the overloaded `tableWrite()` methods, allowing users to specify the compression scheme they want to use when writing the table's output files

Reviewed By: kkaranasos, xiaoxmeng

Differential Revision: D66517547

fbshipit-source-id: 527139263dd8a67fb3d9f7fc470367ca4fcbc619