Add dummy class to fix Nexus release issues
Open source nexus deployment requires the creation of a JAR with Javadoc.  Because we depend
on the code in the `presto-native-execution` module now in other modules for tests, we need
to release these artifacts to Nexus along with the rest of the modules we release.

Because there is only testing code in this module, it fails during Nexus deployment.  Adding
a dummy class with Javadoc to allow this artifact to be released.