Fix initialization order of kSpecialforms (#11770)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11770

Since kSpecialforms is a static, it may be initialized during main before types are actually registered. This can cause failures in the fuzzer. For example, ipaddress types is a custom type that needs to be register before kSpecialforms is actually created.

What we actually want is to use init on use pattern to avoid this issue. We basically had static initialization order issues.

Reviewed By: kagamiori

Differential Revision: D66856845

fbshipit-source-id: 7aa3e7e23f466908403ffb244bf0048bb91a7eb9