fix: Incorrect result when hash probe dynamic filters push down through right join (#11768)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11768

Right/full join could introduce extra null values in the output column compared to probe side, so they should not be considered as "identity projection" and not qualified for dynamic filter push through.  When the condition of completely replacement of hash probe is satisfied, we were allowing all these extra null rows to pass through the join and generating incorrect result.  This change fixes this bug.

It's possible to generate a filter using hash join, but not completely replacing the hash probe though, since the extra nulls should be filtered out during the actual probe process.  Having such mechanism requires extra information to be added in addition to "identity projections" though (i.e. whether only new nulls will be introduced for certain column).  We can revisit this if we see there is a use case for such optimization.

Reviewed By: spershin

Differential Revision: D66833193

fbshipit-source-id: 5e999ca207cddd951793ffbb2ab7ac6fda7c97ae