fix: Possible infinite loop in URIParser's tryConsumeIPV6Address (#11782)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11782

When an IPv6 address in a URI does not end with a closing ']' and this happens at the end of the
string, the parser will enter an infinite loop.

The fix is to break out of the loop when we realize we're at the end of the string.
tryConsumeIPV6Address is not responsible for consume the closing ']' so the caller will realize the
issue when it attempts to consume it (this is already handled properly there).

Reviewed By: kgpai

Differential Revision: D66904149

fbshipit-source-id: 3ad61c68dc19c05f126a4991f8e72303ebc19f94