build: Add option to build monolithic, shared library (#10732)
Summary:
This introduces the option to build a shared version of the monolithic `libvelox.a`, the result is a much smaller build tree, library and executables. They are now so small that moving them is much easier, e.g. for test sharding.

I also fixed some miscellaneous things in this PR:
- Removed `FILESYSTEM`, the link is not necessary with GCC >= 9 which is our minimum.
- Replaced `FOLLYBENCHMARK` with the proper target
- Improved bundled folly's use of glog and gflags (when also bundled)
- [Cleaned up the usage of an addition proto wrapper target](https://github.com/facebookincubator/velox/pull/10732/files#diff-6e1ef8f965a01fe12b35484f26064a6e30207eb3e8722eabe552dba288fbf934R18) so we don't link generated proto files twice (which doesn't cause issues in a static build)
- Added a marker in the log so it's clear when we are running velox cmake vs. dependency cmake:
```
-- [xsimd] xsimd v10.0.0
-- [stemmer] Building stemmer from source
-- Setting Arrow source to AUTO
-- [Arrow] Checking for module 'thrift'
```
Initially I looked at forcing a bundled build of folly and gflags to make sure they are shared but that introduces weird rpath issues when the dependencies are also on the system (like our CI image), instead I just changed the build script to build folly as shared in the image, we maybe want to make that an option so people can keep folly as static (though even for static build shared folly will likely make the binary size much smaller.).  If folly is build from source it will be built static or shared matching the velox build.

I have changed the wrapper function so that only the mono library target is explicitly created as shared, the few other velox utils, test libraries etc. will be build statically to avoid linking and test issues.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/10732

Reviewed By: xiaoxmeng

Differential Revision: D66515203

Pulled By: kgpai

fbshipit-source-id: 260e5c8c9027c560bfc4c0d63a999fa284ef9775