fix: Only register inequality comparisons for orderable types (#11784)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11784

Inequalities only work with types that are orderable pretty much by definition.

For maps in particular we allow users to call map1 < map2 but it will always throw because the map
type does not support inequalities unless null is treated as a value (which it is not in GenericView's
compare function which is what gets invoked).

This will address https://github.com/facebookincubator/velox/issues/11549 where inequalities on
maps can be generated in fuzzers and result in exceptions.

Reviewed By: xiaoxmeng

Differential Revision: D66909744

fbshipit-source-id: cc32dea19939a96a4a01afde050c5664fefb73b3