fix: IN does not handle User errors thrown during eval (#11823)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11823

IN does not handle user errors thrown during eval. This means that if the IN is wrapped with a TRY, user errors
will still get thrown rather than producing a null.

The fuzzer tests demonstrated a case of this, Velox allows creating Timestamps that are large enough that they
cannot be represented as milliseconds, there would be an overflow, in this case toMillis throws a user exception.
When evaluating IN, the argument is converted to milliseconds and compared to the set of Timestamps. When
wrapped in a TRY the IN still throws an exception rather than returning NULL.

Reviewed By: bikramSingh91

Differential Revision: D67063377

fbshipit-source-id: 76638810dd11f7532cf19b754edd2a0f01c18f67