fix(fuzzer): Ensure common dictionary generation skips row_number column (#11821)
Summary:
Ensure that the row_number column, used for lining up results with the
reference database, is not treated as a candidate for wrapping by the
fuzzer's feature that probabilistically adds a common dictionary layer.
This prevents verification mismatches caused by the column being
wrapped and losing its intended purpose.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11821

Reviewed By: kagamiori

Differential Revision: D67071154

Pulled By: bikramSingh91

fbshipit-source-id: 8051d264873fb74bc27064a64027ec72448131ad