fix: json_parse tape failure (#11831)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11831

SIMDJSON fails with a tape failure when if a ':' immediately follows the end of a \" when parsing even though ':' is outside the range of the input. This can happen if the previous string had a ':' at just the right location as we dont clear the buffer when processing strings. Fixes T210222818 .

Reviewed By: amitkdutta

Differential Revision: D67108827

fbshipit-source-id: 6eaee5db4c9d1815a0aa164603ecab8815dd84c2