[native] Enable linux build on master after PR merge matching old CircleCI
The linux-build-and-unit-test CircleCI job also ran after the merge of a PR on
the master branch.
This change enables the same Github action workflow to perform the
same task.

This should also enable caching of master branch build artifacts which
should  help with the build times on first opening ofd a PR as the master branch cache
should be used.