fix(build): Add remote function benchmark under benchmark flags (#11905)
Summary:
The remote function benchmarks are built without checking if the benchmark utilities are being built. These are only built if certain flags are set.

This causes upstream issues in Prestissimo because there remote functions are being built but the benchmarks are not.

PR introducing the breaking change:
https://github.com/facebookincubator/velox/pull/11539

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11905

Reviewed By: kagamiori

Differential Revision: D67403110

Pulled By: kgpai

fbshipit-source-id: 62955ddbc540d4397dce094816496c4d87d035a3