Add support for UPDATE in iceberg
This commit allows users to perform row-level updates when using
the Iceberg connector with Java-based workers.

This is achieved by improving on the IcebergUpdatablePageSource
to implement the updateRows method. The implementation passes
a  generated row ID column as a field in the page required by
updateRows. Then during updateRows, generated a positionDelete
file entry for the row ID, and also writes the row's updated value to a
new page sink for the newly updated data.

These new files are then commited in a rowDelta transaction within
the Iceberg connector metadata after processing is complete.

Co-Authored-By: Nidhin Varghese <Nidhin.Varghese1@ibm.com>
Co-Authored-By: Anoop V S <anoop.v.s@ibm.com>