Add ip_prefix function [4/n] (#11514)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11514

Add ip_prefix function which takes in varchar or ipaddress along with a prefixBits.

Use folly lib to convert the varchar or ipaddress type to folly::IPAddressV6, and then apply the mask.

We return a tuple<int128_t, int8_t> which corresponds to the ipaddress and prefix for IPAddressPrefix type.

Split from https://github.com/facebookincubator/velox/pull/11407

Reviewed By: xiaoxmeng, spershin

Differential Revision: D65802211

fbshipit-source-id: ff0611c7c9aa8fd5057a11d84ecc50ca342ca037