feat(prefix sort): Exclude null byte from sort prefix if column has no nulls (#11583)
Summary:
When inserting data into a RowContainer, we collect information about whether
columns contain null values. If a column does not contain nulls, we can omit
the null byte in the normalized key. This reduces memory usage and decreases
word-by-word comparisons.

For example, with a single sort key of type bigint, without this feature, the
normalized key size is 16 bytes (8 + 1 + alignment padding). With this feature,
the normalized key size is reduced to 8 bytes. However, this improvement only
makes a difference if alignment padding does not negate the optimization. For
instance, with a single sort key of type int, the normalized key size remains 8
bytes regardless of whether this feature is enabled.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11583

Reviewed By: tanjialiang

Differential Revision: D67587505

Pulled By: xiaoxmeng

fbshipit-source-id: 01d7d44d10079f5a4a389518a39eea2ae9c8fab0