feat(fuzzer): Add ability in expression fuzzer to run multiple batches (#11903)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11903

This change adds the ability to run 2 input batches for each
expression fuzzer iteration which will re-use the ExprSet to simulate
its typical usage in actual use-cases like in the ProjectFilter
Operator. The full execution loop of each iteration is modified to
accommodate this change, including input generation and modification,
result verification, re-running input using TRY, finding the minimal
breaking expression tree, and the facility to serialize the input and
repro using the ExpressionRunner utility.

Side note: this exposed a bug in both Simplified and common paths
where the inputs are not cleared if during eval of inputs an exception is
thrown. The fix is also a part of this change.

Reviewed By: kagamiori

Differential Revision: D67368974

fbshipit-source-id: 74984adb9a43cbcd135643c3d5b220eacc207e94