feat: Add cacheable flag in connector split to support fine-grained per-split cache decision (#11977)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11977

Add cacheable flag to support fine-grained per-split cache decision. As Prestissimo might set
cacheable flag in a split (the split context in the protocol) if the split scheduling follow the soft
affinity.

The followup is to add the support in split reader on velox side and set from presto protocol in Prestissimo worker

Reviewed By: zacw7

Differential Revision: D67678409

fbshipit-source-id: f24559b295d1aeb9a9bb1dad849e456fe9a9e41b