refactor: Optimize IndexedPriorityQueue::addOrUpdate 20 times faster (#11955)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/11955

Optimize `IndexedPriorityQueue` to use more compact data structures (binary heap) to leverage cache locality and avoid unnecessary memory allocations.  `addOrUpdate` is now about 20 times faster for large data (~1 million elements).  This allows us to replace the custom heap implementation in `ApproxMostFrequentStreamSummary` with `IndexedPriorityQueue`.

Note that `pop` becomes about twice slower in the new version.  This is expected because we shift some of the reordering cost from insertion time to pop time. This shall be good because for each `pop`, we have at least one corresponding `addOrUpdate` which guarantees net performance gain, and in most of the case we do
not even have `pop` on critical path (e.g. in case of `ApproxMostFrequentStreamSummary`).

Reviewed By: xiaoxmeng

Differential Revision: D67626564

fbshipit-source-id: 8d3543f1b877a134f7c4be5c4405cfc44f90c660