fix(parquet): Support converted type in timestamp reader (#11964)
Summary:
Gluten observed a UT failure at the code location below. The cause is that the
timestamp schema element lacks a logical type and is composed of type, name,
and converted_type. This PR allows parsing converted type from schema elements
and uses it in the timestamp reader to determine the timestamp precision if
logical type is not supplied.

https://github.com/facebookincubator/velox/blob/87c558e094043c5451659c59d38ffc1bd6de781a/velox/dwio/parquet/reader/TimestampColumnReader.h#L99

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11964

Reviewed By: xiaoxmeng

Differential Revision: D67718622

Pulled By: pedroerp

fbshipit-source-id: 948e597447e43444abb2f5f56f17a44fbd0a16ec