feat: [velox][PR] Add Duck plan transformation for qualified joins and subqueries (#11930)
Summary:
DuckDB has a lateral join (DelimJoin) that it introduces into flattened subqueries. This removes the lateral join so that we extract a logical plan that can be used as input to Verax or executed as is.

Adds a way of registering a table to the parser with a type instead of dataset.

This code is strictly a stopgap and will be removed as soon as possible.

TODO: Stop using DuckDB as parser and name resolution.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/11930

Reviewed By: xiaoxmeng

Differential Revision: D67570648

Pulled By: oerling

fbshipit-source-id: 9c8a4c09b945e46644ee2415c51ab8e2de9454c7